Up base shop_oder + up migration + validate lai user va category

+++ database/migrations/2024_10_14_072619_create_shop_oder_items.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedInteger('order_id');
             $table->unsignedInteger('product_id');
             $table->integer('quantity')->default(1);
-            $table->decimal('price', 10, 2);
+            $table->bigInteger('price');
             $table->timestamps();
 
             // Khóa ngoại
